backend/app/services/translation_service.py
import os
from google.cloud import translate_v2 as translate
from typing import Optional

# Ensure GOOGLE_APPLICATION_CREDENTIALS environment variable is set
# to the path of your service account key file.

class TranslationService:
    def __init__(self):
        # self.client = translate.Client() # Initialize client here
        print("TranslationService initialized (mock).")

    async def translate_text(self, text: str, target_language: str = "ur") -> Optional[str]:
        """
        Translates text to the target language using Google Translate API.
        """
        if not text:
            return None

        # Check for Google Credentials
        if not os.getenv("GOOGLE_APPLICATION_CREDENTIALS"):
            print("GOOGLE_APPLICATION_CREDENTIALS not set. Translation skipped.")
            return f"Mock translation for '{text}' to {target_language} (Credentials Missing)"

        try:
            # For now, return a mock response
            # In real implementation:
            # result = self.client.translate(text, target_language=target_language)
            # return result['translatedText']
            return f"Mock translation for '{text}' to {target_language} (API Call Simulated)"
        except Exception as e:
            print(f"Error during translation: {e}")
            return None

# Singleton instance of the service
translation_service = TranslationService()
